+++ code/test/sixsq/nuvla/server/resources/callback_lifecycle_test.clj
@@ -163,7 +163,6 @@
 (deftest bad-methods
   (let [resource-uri (str p/service-context (u/new-resource-id callback/resource-type))]
     (ltu/verify-405-status [[base-uri :options]
-                            [base-uri :delete]
                             [resource-uri :options]
                             [resource-uri :put]
                             [resource-uri :post]])))