+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Unreleased
 
 ### Change
-
+  - Support bulk delete for more resources and fix operations for collection
   - Job - Stale job because of race condition fix
   - NuvlaBox - Simplify names of nuvlabox sub-resources at commission time